Dialog Box: initial commit

+++ src/app/components/editor/audiotracklist.tsx
@@ -10,6 +10,7 @@ import { FaCopy, FaTrash } from "react-icons/fa";
 import { removeAudioFromAllTracks } from "@/app/state/trackdetails";
 import { audioManager } from "@/app/services/audiotrackmanager";
 import { deleteColor } from "@/app/services/color";
+import { DialogContext } from "@/app/providers/dialog";
 
 export function AudioTrackList() {
   // Selectors
@@ -22,6 +23,11 @@ export function AudioTrackList() {
     // isContextOpen,
     showContextMenu
   } = React.useContext(ContextMenuContext);
+
+  const {
+    showDialog,
+    hideDialog
+  } = React.useContext(DialogContext);
  
   const dispatch = useDispatch();
 
@@ -80,8 +86,19 @@ export function AudioTrackList() {
         name: 'Delete',
         icon: <FaTrash />,
         onSelect: () => {
-          deleteTrack(index);
-          hideContextMenu();
+          showDialog({
+            confirm: () => {
+              deleteTrack(index);
+              hideDialog();
+              hideContextMenu();
+            },
+            cancel: () => {
+              hideDialog();
+              hideContextMenu();
+            },
+            message: `Are you sure to delete this track?`,
+            messageHeader: <h1 className="text-xl">Confirm Delete Track <b>"{files[index].audioName}"</b></h1>
+          })
         },
       },
     ], event.nativeEvent.clientX, event.nativeEvent.clientY);
@@ -97,7 +114,7 @@ export function AudioTrackList() {
           Load Audio
         </button>
       </div>
-      <div className="list">
+      <div className="list w-full h-[80dvh] overflow-y-scroll">
         {files.map((file: AudioDetails, index: number) => {
           const isSame = selected.audioId === file.audioId;
